General refactoring + renaming of TimingSectionContainer.
