+++ osu.Game.Rulesets.Mania/osu.Game.Rulesets.Mania.csproj
@@ -70,7 +70,7 @@
     <Compile Include="ManiaRuleset.cs" />
     <Compile Include="Mods\ManiaMod.cs" />
     <Compile Include="UI\SpecialColumnPosition.cs" />
-    <Compile Include="UI\TimingSectionContainer.cs" />
+    <Compile Include="Timing\TimeRelativeContainer.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\osu-framework\osu.Framework\osu.Framework.csproj">
